feat: add code for esm fn
